Moved DTO classes to proper package
