+++ src/main/java/com/idatt2105/backend/controller/QuestionController.java
@@ -1,9 +1,9 @@
 package com.idatt2105.backend.controller;
 
 import com.idatt2105.backend.model.Alternative;
-import com.idatt2105.backend.model.AlternativeDTO;
+import com.idatt2105.backend.dto.AlternativeDTO;
 import com.idatt2105.backend.model.Question;
-import com.idatt2105.backend.model.QuestionDTO;
+import com.idatt2105.backend.dto.QuestionDTO;
 import com.idatt2105.backend.model.Tag;
 import com.idatt2105.backend.service.QuestionService;
 import io.swagger.v3.oas.annotations.Operation;
