+++ src/main/java/com/idatt2105/backend/model/MultipleChoiceQuestion.java
@@ -1,11 +1,11 @@
 package com.idatt2105.backend.model;
 
+import com.idatt2105.backend.dto.AlternativeDTO;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import java.util.HashSet;
-import java.util.Objects;
 import java.util.Set;
 import lombok.Data;
