+++ src/main/java/com/idatt2105/backend/service/QuestionService.java
@@ -1,10 +1,10 @@
 package com.idatt2105.backend.service;
 
 import com.idatt2105.backend.model.Alternative;
-import com.idatt2105.backend.model.AlternativeDTO;
+import com.idatt2105.backend.dto.AlternativeDTO;
 import com.idatt2105.backend.model.MultipleChoiceQuestion;
 import com.idatt2105.backend.model.Question;
-import com.idatt2105.backend.model.QuestionDTO;
+import com.idatt2105.backend.dto.QuestionDTO;
 import com.idatt2105.backend.model.Quiz;
 import com.idatt2105.backend.model.Tag;
 import com.idatt2105.backend.model.TrueOrFalseQuestion;
@@ -14,10 +14,7 @@ import com.idatt2105.backend.repository.QuestionRepository;
 import com.idatt2105.backend.repository.TagRepository;
 import com.idatt2105.backend.util.InvalidIdException;
 import jakarta.validation.constraints.NotNull;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
-import java.util.Set;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
